Project/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
python-dotenv = "^0.21.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^4.4.2"
fastapi-limiter = "^0.1.5"
python-multipart = "^0.0.5"
email-validator = "^1.3.0"
fastapi-mail = "1.2.2"
httpx = "^0.23.1"
requests = "^2.28.1"
aerich = "^0.7.1"
fastapi-sso = "^0.6.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = {extras = ["toml"], version = "^6.5.0"}
black = "^22.10.0"


[tool.black]
target-version = ['py311']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''


[tool.coverage]
    [tool.coverage.run]

    source = ["source"]
    branch = true
    omit = ['']

    [tool.coverage.report]
    exclude_lines = [
        "# noqa",
        "raise NotImplementedError"
    ]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
